Modify DIRS and APP_DIRS in TEMPLATES
By adding the templates' paths to DIRS, we define a list of directories
the engine should look for HTML source files.

Changing the value of APP_DIRS from True to False makes the engine
not to look for templates inside installed applications. It's changed
to False because there is only one directory for storing templates
(blog/templates).